- на евенты трупы больше не залетают - мелуха по custom htm
git-svn-id: https://subversion.assembla.com/svn/scoria4@29 0f1168a5-6793-4626-a6db-a0a741385511

+++ interlude/core/GameServer/java/com/l2scoria/loginserver/network/serverpackets/ServerList.java
@@ -136,7 +136,7 @@ public final class ServerList extends L2LoginServerPacket
 			writeC(server._pvp ? 0x01 : 0x00);
                         if(Config.FAKE_SERVER_LIST) {
                             if(Config.FAKE_SERVER_LIST_TYPE.equals("STATIC")) {
-                                writeH(Config.FAKE_SERVER_LIST_PARAM1);
+                                writeH((int)Config.FAKE_SERVER_LIST_PARAM1);
                                 writeH(Config.FAKE_SERVER_LIST_PARAM2);
                             } else if(Config.FAKE_SERVER_LIST_TYPE.equals("DYNAMIC")) {
                                 double mulOpertype = server._currentPlayers * Config.FAKE_SERVER_LIST_PARAM1;
